Replace my hand-rolled parser with `bencoding`
It was nice to write a simple parser with `attoparsec` and I though it
was wonderful. Until I had to wrap/unwrap ADTs with it, at which point I
realized I'd have to write my own classes and recursive
serialization/deserialization + make it efficient.

I would never do this for JSON parsing and I won't do it for this
protocol. It's sad and for all other libraries apart from this one
(`bencoding`) it really wouldn't be worth the cost to include a new
dependency. But this is hand optimized and well-written. The code is
about 1000 lines of manual good parsing and unparsing, made to be fast
and generalized.

I currently lack the expertise to implement a full-fledged parsing
library that's worth the time.